+++ database/migrations/2021_12_10_1639179894_create_revenue_fules_table.php
@@ -16,9 +16,9 @@ class CreateRevenueFulesTable extends Migration
     {
         Schema::create('revenue_fules', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigInteger('quantity');
-            $table->bigInteger('price');
-            $table->bigInteger('paid_amount')->nullable();
+            $table->double('quantity');
+            $table->double('price');
+            $table->double('paid_amount')->nullable();
             $table->foreignId("filling_id")->constrained("fillings")->references("id");
             $table->foreignId("stock_id")->constrained("stocks")->references("id");
             $table->foreignId("revenue_id")->constrained("revenues")->references("id");
